refactor: mutualized scattered types
