+++ .eslintrc
@@ -4,6 +4,7 @@
     "arrow-parens": "off",
     "semi": ["error", "always"],
     "space-before-function-paren": ["error", "never"],
-    "no-template-curly-in-string": "off"
+    "no-template-curly-in-string": "off",
+    "vue/multi-word-component-names": "off"
   }
 }
